Add files via upload
Improved conversion to UD of the original Index following the errors detected by the new validation script. Various errors in the original annotation have been corrected in the process.

Note: this conversion still does not pass completely the validation test, especially because of non-projective punctuation, but we have to defer a more thourough improvement to a later date.